Remove default for TAGS, update comments
The TAGS default was incorrect (CSV, should have been space separated).
Better to remove the default for tags so the user is not surprised when
running a job with only FINALLY_TAGS.

Comments are updated, and write_properties is no longer called before
tags, as claim is now a tag, so no further information can be known
before at least one tag as run.