Add ssh_retry plugin
Retry failed ssh connections as soon as they happen. Some ssh
connections fail, not sure why. Its better to retry them there and then
rather than failing out the whole play(book) and having to start agian
from the beginning.

This plugin was removed from OSAD due to a licencing conflict.
A version of it has been accepted upstream into ansible v2, but until
that is released we must carry the patch for v1.x if we want to be able
to retry failed tasks.